fix(nika-schema): rename private module — rustdoc name collision
main carried both 'mod infer_permits' and 'pub fn infer_permits' in
check/mod.rs — legal Rust (type vs value namespace), but rustdoc
--document-private-items refuses to emit pages for both and hygiene
vector 28 (doc-private-items, ADR-015 Gate-8-extended) went RED on
main, blocking every push. Cross-PR class: each side landed green on
its own base; the combination is what broke.

The module is private, so the rename is invisible outside the crate:
permits_infer now mirrors its sibling permits_fit (one verifies the
declared boundary, one infers the tightest boundary). The public
surface — check::infer_permits, check::task_permits,
check::InferredPermits — is byte-identical (public-api.txt untouched).

Proof: RUSTDOCFLAGS='-D warnings' cargo doc --no-deps
--document-private-items -p nika-schema GREEN · clippy 0 · 735 lib
tests pass.
