+++ crates/nika-schema/src/check/permits_infer.rs
